requirements.txt
datasets>=2.14.0
pandas>=2.0.0
pytest>=7.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

